+++ src/test/java/net/minestom/server/instance/EntityTrackerTest.java
@@ -1,4 +1,4 @@
-package instance;
+package net.minestom.server.instance;
 
 import net.minestom.server.coordinate.Vec;
 import net.minestom.server.entity.Entity;
