+++ src/test/java/net/minestom/server/instance/PaletteTest.java
@@ -1,4 +1,4 @@
-package instance;
+package net.minestom.server.instance;
 
 import net.minestom.server.instance.palette.Palette;
 import org.junit.jupiter.api.Test;
