+++ src/test/java/net/minestom/server/utils/ObjectArrayTest.java
@@ -1,6 +1,5 @@
-package utils;
+package net.minestom.server.utils;
 
-import net.minestom.server.utils.ObjectArray;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
